% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocksMath.R
\name{calculate_TTD}
\alias{calculate_TTD}
\title{Transactional times duraion}
\usage{
calculate_TTD(df, delete_overnight)
}
\arguments{
\item{df}{dataframe with columns: Date, Time}

\item{delete_overnigh}{pameter TRUE or FALSE}
}
\value{
dataframe with Date, Time, TTD - "calculated Transactianal times duration"
}
\description{
This function lets you to calculate transactional times duration
i.e. the interval between consecutive transactions in seconds,
function also can delete overnight TTD
}
