% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocksMath.R
\name{EWMA_volatility}
\alias{EWMA_volatility}
\title{EWMA volatility- exponentially weighted moving average of volatility}
\usage{
EWMA_volatility(returns, lambda)
}
\arguments{
\item{returns}{Vector with stock returns.}

\item{lambda}{Lambda parameter}
}
\value{
A vector of calculated EWMA volatility.
}
\description{
This function lets you to calculate EWMA_volatility -
exponentially weighted moving average of variance.
}
