% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocksMath.R
\name{log_returns}
\alias{log_returns}
\title{Logarithmic returns}
\usage{
log_returns(vector)
}
\arguments{
\item{vector}{Vector with prices of the company.}
}
\value{
A vector with the logarithmic return of investment.
}
\description{
This function lets you to calculate the logarithmic return of investment.
}
