% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocksMath.R
\name{mean_ttd}
\alias{mean_ttd}
\title{Calculate mean Transactional times duraion for given time interval}
\usage{
mean_ttd(df, period)
}
\arguments{
\item{df}{dataframe with columns: Date, Time, TTD}
}
\value{
dataframe with Date, Time, mean_TTD -
"calculated mean Transactianal times duration for each interval"
}
\description{
This function lets you to calculate mean transactional times duration
for given time interval
}
