% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocksMath.R
\name{moving_pearson}
\alias{moving_pearson}
\title{Moving Pearson correlation}
\usage{
moving_pearson(vector1, vector2, window)
}
\arguments{
\item{vector1}{Vector of first series.}

\item{vector2}{Vector of second series.}

\item{time_window}{Time window.}
}
\value{
Vector of calculated moving Pearson correlation.
}
\description{
This function lets you to calculate moving Pearson correlation
between two series with given time window.
}
