% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocksMath.R
\name{simple_returns}
\alias{simple_returns}
\title{Simple return of investment}
\usage{
simple_returns(vector)
}
\arguments{
\item{vector}{Vector with prices of the company.}
}
\value{
A vector with the simple return of investment.
}
\description{
This function lets you to calculate the simple return of investment.
}
