% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocksMath.R
\name{weighted_prices}
\alias{weighted_prices}
\title{Calculate weighted arithmetic mean prices for intraday stocks transactions
where volume in weight}
\usage{
weighted_prices(intradayTransactions)
}
\arguments{
\item{intradayTransactions}{dataframe with columns: Date, Time, Price, Volume}
}
\value{
dataframe with Date, Time, Price
}
\description{
This function lets you to calculate weighted arithmetic
mean prices for intraday stocks transactions for each
unique date and hour where volume is weight
}
